refactor: adjust folder for mcp connection
